+++ docs/changes.rst
@@ -10,6 +10,7 @@ Change Log
   * added :meth:`.get_r` to recursively copy remote directories to a local path
   * added :attr:`.pwd` to return the current working directory
   * added :meth:`.cwd` as synonym for :meth:`.chdir`
+  * modified :meth:`.listdir` to return a sorted list instead of an arbitrary one
 
 
 * 0.2.6 (current, released 2014-05-17)